test: platform-policy test harness improvements
Two improvements for testing platform behavior (supautils policy against
the rendered config and real roles):

1. Split testinfra/test_ami_nix.py (1400+ lines) into domain modules
   sharing the session-scoped EC2 fixture via conftest.py:
   test_postgrest, test_postgres, test_jit_pam, test_overrides,
   test_mitigations, test_apparmor. The AMI suite stays for artifact
   wiring checks; pytest now runs the whole directory.

2. Document (nix/tests/README.md) that the pg_regress suite already runs
   against the rendered supautils.conf.j2 with migrations applied via
   start-postgres-server, and establish the SET ROLE convention for
   platform-policy tests, with supautils_platform_policy.sql as the
   reference example (privileged-extension delegation as the
   non-superuser postgres role).

Conf changes and their policy tests can now travel in the same PR.
